refactor: 도메인 기반 폴더 구조 정리
