+++ src/app/(main)/(owner)/[id]/@modal/menu/category/add/layout.tsx
@@ -1,6 +1,6 @@
 import { PropsWithChildren } from "react";
 import getQueryClient from "@/app/get-query-client";
-import { getCategories } from "@/lib/api/menu.api";
+import { getCategories } from "@/app/(main)/(owner)/[id]/menu/_api/menu.api";
 import { dehydrate, HydrationBoundary } from "@tanstack/react-query";
 import { categoryKeys } from "../../../../menu/_queries/keys";
 import ClientRefWrapper from "../../_components/Wrapper";
