+++ src/app/(main)/(owner)/[id]/device/layout.tsx
@@ -1,6 +1,6 @@
 import PageTitle from "@/app/(main)/_components/PageTitle";
 import getQueryClient from "@/app/get-query-client";
-import { getDevices } from "@/lib/api/device.api";
+import { getDevices } from "@/app/device/_api/device.api";
 import { dehydrate, HydrationBoundary } from "@tanstack/react-query";
 import { PropsWithChildren } from "react";
