+++ src/app/(main)/(owner)/[id]/menu/_queries/useMenu.tsx
@@ -9,7 +9,7 @@ import {
   postMenu,
   updateMenuWithImage,
   updateMenuWithoutImage,
-} from "@/lib/api/menu.api";
+} from "@/app/(main)/(owner)/[id]/menu/_api/menu.api";
 import { useMutation, useQuery } from "@tanstack/react-query";
 import { menuKeys } from "./keys";
