+++ src/app/(main)/(owner)/[id]/settings/layout.tsx
@@ -1,6 +1,6 @@
 import PageTitle from "@/app/(main)/_components/PageTitle";
 import getQueryClient from "@/app/get-query-client";
-import { getStoreInfoDetail } from "@/lib/api/stores.api";
+import { getStoreInfoDetail } from "@/app/(main)/(owner)/[id]/store/_api/stores.api";
 import { dehydrate, HydrationBoundary } from "@tanstack/react-query";
 import { PropsWithChildren } from "react";
 import { settingsKeys } from "./_queries/keys";
