+++ src/app/waiting/_queries/keys.ts
@@ -1,0 +1,3 @@
+export const waitingKeys = {
+  all: () => ["waiting-list"] as const,
+};
