+++ src/app/waiting/page.tsx
@@ -3,7 +3,7 @@
 import ResponsiveButton from "@/components/common/Button/ResponsiveButton";
 import { useRouter } from "next/navigation";
 import WaitingSection from "./_components/WaitingSection";
-import useWaiting from "./_hooks/useWaiting";
+import useWaiting from "./_queries/useWaiting";
 import useDeviceInfo from "./_hooks/useDeviceInfo";
 import useWaitingModal from "./_hooks/useWaitingModal";
