+++ tsconfig.json
@@ -28,7 +28,7 @@
   },
   "include": [
     "src",
-    "src/app/**/types/*.d.ts",
+    "src/app/**/_types/*.d.ts",
     "src/types/index.d.ts",
     "tests",
     "next-env.d.ts",
